small refactor to simplify layout
